Commit de Ronan Le Boulch : Correction du bug pour MPR et MPF concernant l'OS Linux en 64 bits. Le format utilisé pour récupérer la position et la longueur est un uint32 quelque soit l'architecture
git-svn-id: svn+ssh://scm.gforge.inria.fr/svnroot/mptk/trunk@1465 0f77665a-23fb-0310-bc23-e91fd8efc3e1

+++ src/libmptk/atom.in.cpp
@@ -171,8 +171,8 @@ int MP_Atom_c::read( FILE *fid, const char mode ) {
   case MP_BINARY:
     /* Support */
     for ( i=0, nItem = 0; i<numChans; i++ ) {
-      nItem += mp_fread( &(support[i].pos), sizeof(unsigned long int), 1, fid );
-      nItem += mp_fread( &(support[i].len), sizeof(unsigned long int), 1, fid );
+      nItem += mp_fread( &(support[i].pos), sizeof(uint32_t), 1, fid );
+      nItem += mp_fread( &(support[i].len), sizeof(uint32_t), 1, fid );
     }
     /* Amp */
     if ( mp_fread( amp,   sizeof(MP_Real_t), numChans, fid ) != (size_t)numChans ) {
